feat: build up submission page I've added a textarea and a place to render the sketch
