+++ ui/src/pages/Submit/Submit.css
@@ -1,38 +1,29 @@
-.App {
+.submit-page {
   text-align: center;
 }
 
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
+.sketch-container {
+  display: inline-grid;
+  grid-template-columns: 50% 50%;
+  grid-column-gap: 20px;
+  width: 90%;
+  margin: 0 5%;
 }
 
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
+.sketch-code {
+  min-height: 50vh;
 }
 
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
+.sketch-preview {
+  border: 1px solid #000;
 }
 
-.App-link {
-  color: #61dafb;
+.sketch-submit {
+  width: 300px;
+  padding: 20px 0;
+  font-size: 25px;
+  width: 30%;
+  margin: 0 auto;
+  cursor: pointer;
 }
 
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
